requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
line-bot-sdk==3.14.2
gspread==6.1.4
google-auth==2.37.0
pytest==8.3.4
tzdata==2025.2
